requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
qdrant-client
python-dotenv
